:sparkles: Add Obj Model

+++ core/narc_engine/include/models/Mesh.h
@@ -1,19 +1,18 @@
 #pragma once
 
-namespace narc_engine
-{
+namespace narc_engine {
     struct Vertex;
 
-    template <typename T>
+    template<typename T>
     class GraphicsBuffer;
 
     typedef GraphicsBuffer<Vertex> VertexBuffer;
-    typedef GraphicsBuffer<uint16_t> IndexBuffer;
+    typedef GraphicsBuffer<uint32_t> IndexBuffer;
 
     class NARC_ENGINE_API Mesh
     {
     public:
-        Mesh(const std::vector<Vertex>& vertices, const std::vector<uint16_t>& indices);
+        Mesh(const std::vector<Vertex>& vertices, const std::vector<uint32_t>& indices);
         ~Mesh();
 
         const VertexBuffer* getVertexBuffer() const { return m_vertexBuffer.get(); }
